+++ kaithem/src/version_info.py
@@ -1,2 +1,2 @@
-__version__ = "0.68.39 Production"
-__version_info__ = (0, 68, 39, "release", 0)
+__version__ = "0.68.42 Production"
+__version_info__ = (0, 68, 42, "release", 0)